COS30049_RazorChain-main/COS30049_RazorChain_Api/API/app/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
mysqlclient
pydantic[email]